pyproject.toml
[project]
name = "dl-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "python-dotenv>=1.2.1",
    "pytorch-lightning>=2.6.0",
    "rich>=14.3.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "wandb>=0.24.0",
]
